Bachelorarbeit-XAI-PM/requirements.txt
# Zentrale Data-Science-Bibliotheken
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Explainable-AI-Bibliotheken
shap==0.44.0

# Visualisierung
matplotlib==3.8.2
seaborn==0.13.0
graphviz==0.21 

# Machine-Learning-Modelle
xgboost==2.0.3

# Hilfsbibliotheken
pyyaml==6.0.1
joblib==1.3.2

# Jupyter (für interaktive Notebooks)
jupyter==1.0.0
ipykernel==6.27.1